% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankCustomer.R
\docType{data}
\name{BankCustomer}
\alias{BankCustomer}
\title{Customers of a bank}
\format{
A data frame with 150 rows and 9 variables:
\describe{
\item{age}{age of the customer}
\item{anciennete}{Since when the customer is in the bank}
\item{profession}{work of the customer}
\item{revenu}{How much the customer earn}
\item{epargne}{How much the customer saving}
\item{carte_bleue}{If the customer get a credit card}
\item{pea}{action savings plan}
\item{score}{score attribuate by the bank}
\item{Cluster}{Cluster attribuate by the bank}
}
}
\usage{
BankCustomer
}
\description{
Customers of a bank
}
\keyword{datasets}
