% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{CalcTable.Data}
\alias{CalcTable.Data}
\title{CalcTable.Data}
\usage{
CalcTable.Data(object, varqual)
}
\arguments{
\item{object}{a Dataset object}

\item{varqual}{the other variable to cross}
}
\value{

}
\description{
CalcTable.Data
}
