% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateFactor.R
\name{contingency}
\alias{contingency}
\title{Contingency table}
\usage{
contingency(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active categorical variable}
}
\value{
A cross table between two variables.
The same table with margins and sum
The table with percentages by lines
Number of lines (i.e. number of clusters)
Number of columns (i.e. number of modes in the other variables)
}
\description{
Create a cross table between a cluster variable and a selected categorical variable.
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
table <- contingency(obj, BankCustomer$profession)
effectiveTable <- table[[2]]
}
