% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{Corr_ratios.NumDataset}
\alias{Corr_ratios.NumDataset}
\title{Corr_ratios.NumDataset}
\usage{
Corr_ratios.NumDataset(object)
}
\arguments{
\item{object}{a NumDataset object}
}
\value{

}
\description{
Corr_ratios.NumDataset
}
