% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{davies_bouldinC}
\alias{davies_bouldinC}
\title{Davies-Bouldin Index}
\usage{
davies_bouldinC(object, clusters = object$pred_clusters)
}
\arguments{
\item{object}{An object of class ccdata}

\item{clusters}{A vector corresponding to the dataset clustering results}
}
\value{
The Davies-Bouldin index for all the cluster groups.
}
\description{
Davies-Bouldin Index
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
davies_bouldinC(obj)
}
