% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{dunn_indexC}
\alias{dunn_indexC}
\title{Dunn Index}
\usage{
dunn_indexC(object, clusters = object$pred_clusters)
}
\arguments{
\item{object}{An object of class ccdata}

\item{clusters}{A vector corresponding to the dataset clustering results (predicted clusters with ccdata class object by default)}
}
\value{
The Dunn Index for all the cluster groups
}
\description{
Dunn Index
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
dunn_indexC(obj)
}
