% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{effectsize}
\alias{effectsize}
\title{Effect Size for numerical variables}
\usage{
effectsize(object)
}
\arguments{
\item{object}{an object of class ccdata}
}
\value{
An effect size table
matrix representing the effect size of every variable for each cluster.
The rows represent the variables. The columns represent the clusters.
}
\description{
Effect Size for numerical variables
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
effectsize(obj)
}
