% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{EffectSizeTable.NumDataset}
\alias{EffectSizeTable.NumDataset}
\title{EffectSizeTable.NumDataset}
\usage{
EffectSizeTable.NumDataset(object)
}
\arguments{
\item{object}{a NumDataset object}
}
\value{

}
\description{
EffectSizeTable.NumDataset
}
