% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{FactorDataset}
\alias{FactorDataset}
\title{FactorDataset}
\usage{
FactorDataset(data, vargroupe)
}
\arguments{
\item{data}{a dataset with factor variables}

\item{vargroupe}{the variable's name of the cluster}
}
\value{
a Factordataset object
}
\description{
FactorDataset
}
