% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{get_FAMD}
\alias{get_FAMD}
\title{FAMD}
\usage{
get_FAMD(object)
}
\arguments{
\item{object}{An object of class ccdata
Show three graphs to illustrate the cluster with all the variables of your dataset.}
}
\value{
B. The plot with all the modalities of the category variables placed by importance of contribution with the two first axes.
C. Correlation circle with all the numeric variables
A. Plot with all the individuals colored by their cluster group and illustrate by category and numeric variables
}
\description{
FAMD
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
get_FAMD(obj)
}
