% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{Get_MCA.Data}
\alias{Get_MCA.Data}
\title{Get_MCA.Data}
\usage{
Get_MCA.Data(object, index_names)
}
\arguments{
\item{object}{a data object}

\item{index_names}{optional vector of the index to use}
}
\value{

}
\description{
Get_MCA.Data
}
