% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentAnalysis.R
\name{get_MCA}
\alias{get_MCA}
\title{Multiple Component Analysis
Show two graphs to illustrate the cluster with all the categorical variables of your dataset.}
\usage{
get_MCA(object)
}
\arguments{
\item{object}{An object of class ccdata}
}
\value{
A first plot with all the modes of the categorical variables placed by importance of contribution with the two first axes.
A biplot whho illustrate the cluster group with all the modes of categorical variables.
}
\description{
Multiple Component Analysis
Show two graphs to illustrate the cluster with all the categorical variables of your dataset.
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
get_MCA(obj)
}
