% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{Get_PCA.Data}
\alias{Get_PCA.Data}
\title{Get_PCA.Data}
\usage{
Get_PCA.Data(object, index_names)

Get_PCA.Data(object, index_names)
}
\arguments{
\item{object}{a data object}

\item{index_names}{optional vector of the index to use}
}
\value{



}
\description{
Get_PCA.Data

Get_PCA.Data
}
