% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{get_PCA}
\alias{get_PCA}
\title{get_PCA}
\usage{
get_PCA(object, index_names = NULL)
}
\arguments{
\item{object}{an object of class ccdata}

\item{index_names}{optional vector of the index to use}
}
\value{

}
\description{
get_PCA
}
