% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{NumDataset}
\alias{NumDataset}
\title{NumDataset}
\usage{
NumDataset(data, clusters_data)
}
\arguments{
\item{data}{a dataset}

\item{clusters_data}{a Numdataset object}
}
\value{

}
\description{
NumDataset
}
