% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{PhiValueTable.FactorDataset}
\alias{PhiValueTable.FactorDataset}
\title{PhiValueTable.FactorDataset}
\usage{
PhiValueTable.FactorDataset(object, nomvarqual)
}
\arguments{
\item{object}{a object}

\item{nomvarqual}{a string}
}
\value{

}
\description{
PhiValueTable.FactorDataset
}
