% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor.R
\name{PhiValueTable.MultiFactor}
\alias{PhiValueTable.MultiFactor}
\title{PhiValueTable.MultiFactor}
\usage{
PhiValueTable.MultiFactor(object, nomvarquali)
}
\arguments{
\item{object}{a MultiFactor object}

\item{nomvarquali}{a string}
}
\value{

}
\description{
PhiValueTable.MultiFactor
}
