% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotphi}
\alias{plotphi}
\title{Title}
\usage{
plotphi(obj, var)
}
\arguments{
\item{obj}{an object of class Data}

\item{var}{name of variable of interest to plot phi-values}
}
\value{

}
\description{
Title
}
