% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plottvalue}
\alias{plottvalue}
\title{Plot of t-values}
\usage{
plottvalue(object, var = NULL)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active variable}
}
\value{
plot of the tests value for all numerical variables and all clusters or plot of tests values by modes of a selected categorical variable for all clusters.
}
\description{
Plot of t-values
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
plottvalue(obj) #for all the numerical variables
plottvalue(obj, BankCustomer$profession) #for only one categorical variable
}
