% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plottvaluefactor}
\alias{plottvaluefactor}
\title{Title}
\usage{
plottvaluefactor(obj, var)
}
\arguments{
\item{obj}{an object of class FactorDataset}

\item{var}{name of variable of interest to plot t-values}
}
\value{

}
\description{
Title
}
