% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hierarchical.R
\name{run_hierarchical}
\alias{run_hierarchical}
\title{Hierarchical Clustering for a vector of Cluster Sizes}
\usage{
run_hierarchical(data_input, cluster_sizes, method = "ward.D", members = NULL)
}
\arguments{
\item{data_input}{A distance matrix. Can be one of the distance objects from \code{\link[=dimension_reduction_list]{dimension_reduction_list()}}.}

\item{cluster_sizes}{a vector of cluster sizes.
The value is used as the number of centers within \code{\link[stats:kmeans]{stats::kmeans()}}.}

\item{method}{the agglomeration method to be used.  This should
    be (an unambiguous abbreviation of) one of
    \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
    \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA),
    \code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{members}{\code{NULL} or a vector with length size of
    \code{d}. See the \sQuote{Details} section.}
}
\value{
a named list containing classification vectors of the number
of cluster sizes provided.
}
\description{
A function that runs hclust from \code{\link[stats:hclust]{stats::hclust()}} through a loop using the dimension reduction taken from \code{\link[=dimension_reduction_list]{dimension_reduction_list()}}.
}
\examples{
# load base iris dataset from rstudio
data("iris")
x <- t(iris[,1:4])
y <- dimension_reduction_list(x)

# for a single data structure
run_hierarchical(y$pca_distance_object, 1:5)

}
\seealso{
\code{\link[stats:hclust]{stats::hclust()}}
}
