% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_leiden.R
\name{run_leiden}
\alias{run_leiden}
\title{Leiden Clustering for a vectors of Parameters}
\usage{
run_leiden(data_input, k_nearest_neighbors, resolution)
}
\arguments{
\item{data_input}{Either a diagonal numeric matrix or a distance object from \code{\link[stats:dist]{stats::dist()}}. The distance_object output from \code{dimension_reduction_list()}}

\item{k_nearest_neighbors}{number of neighbors to be used when constructing the nearest neighbor graph.}

\item{resolution}{A scalar between 0 and 1. The \code{resolution_parameter} from \code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}.}
}
\value{
a named list containing classification vectors of the parameters provided.
}
\description{
conducts Leiden clustering for multiple vectors of parameter choices.
}
\examples{
# load base iris dataset from rstudio
data("iris")
x <- t(iris[,1:4])
y <- dimension_reduction_list(x)

run_leiden(y$pca_distance_object, c(20, 25), c(0.5, 0.8))

}
\seealso{
\code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}, \code{\link[cccd:nng]{cccd::nng()}}
}
