% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_louvain.R
\name{run_louvain}
\alias{run_louvain}
\title{Louvain Clustering for a vectors of Parameters}
\usage{
run_louvain(data_input, k_nearest_neighbors, resolution)
}
\arguments{
\item{data_input}{Either a diagonal numeric matrix or a distance object from \code{\link[stats:dist]{stats::dist()}}. The distance_object output from \code{dimension_reduction_list()}}

\item{k_nearest_neighbors}{number of neighbors to be used when constructing the nearest neighbor graph.}

\item{resolution}{A scalar between 0 and 1. The \code{resolution} parameter from \code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}.}
}
\value{
a named list containing classification vectors of the parameters provided.
}
\description{
conducts Louvain clustering for multiple vectors of parameter choices.
}
\examples{
# load base iris dataset from rstudio
data("iris")
x <- t(iris[,1:4])
y <- dimension_reduction_list(x)

run_louvain(y$pca_distance_object, c(20, 25), c(0.5, 0.8))

}
\seealso{
\code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}, \code{\link[cccd:nng]{cccd::nng()}}
}
