% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{silhouette.Data}
\alias{silhouette.Data}
\title{silhouette.Data}
\usage{
silhouette.Data(object, Clusters)
}
\arguments{
\item{object}{a datset object}

\item{Clusters}{a vector cluster}
}
\value{

}
\description{
silhouette.Data
}
