% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{silhouette}
\alias{silhouette}
\title{silhouette}
\usage{
silhouette(data, clusters)
}
\arguments{
\item{clusters}{a vector corresponding to the dataset clustering results}

\item{object}{an object of class ccdata}
}
\value{

}
\description{
silhouette
}
