% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{silhouetteC}
\alias{silhouetteC}
\title{Silhouette}
\usage{
silhouetteC(object, clusters = object$pred_clusters)
}
\arguments{
\item{object}{An object of class ccdata}

\item{clusters}{A vector corresponding to the dataset clustering results. Automatically equal of the cluster vector entered as a Dataset parameter.}
}
\value{
The silouhette value for all the cluster groups and a mean silouhette.
}
\description{
Silhouette
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
silhouetteC(obj)
}
