% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{TestStatistique.Data}
\alias{TestStatistique.Data}
\title{TestStatistique.Data}
\usage{
TestStatistique.Data(object, varexp)
}
\arguments{
\item{object}{a data object}

\item{varexp}{a string}
}
\value{

}
\description{
TestStatistique.Data
}
