% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{Transformdata.Data}
\alias{Transformdata.Data}
\title{Transformdata.Data}
\usage{
Transformdata.Data(object)
}
\arguments{
\item{object}{a data object}
}
\value{

}
\description{
Transformdata.Data
}
