% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{transformdata}
\alias{transformdata}
\title{transformdata}
\usage{
transformdata(object)
}
\arguments{
\item{object}{An object of class ccdata}
}
\value{
A new dataset for the metrics function, created with the new coordinates based on a FAMD.
}
\description{
Data tranformation for categorical and mixed variables
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
transformdata(obj)
}
