% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{tvalue_cat}
\alias{tvalue_cat}
\title{t-value for categorical variables}
\usage{
tvalue_cat(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active categorical variable}
}
\value{
A cross table between the cluster variable and the selected variable with test values for all cluster*modality pairs.
}
\description{
Calculates the test values for all cluster*modes of one selected variable.
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
tvalue_cat(obj, BankCustomer$profession)
}
