% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{TValueTable.FactorDataset}
\alias{TValueTable.FactorDataset}
\title{TValueTable.FactorDataset}
\usage{
TValueTable.FactorDataset(object, nomvarqual)
}
\arguments{
\item{object}{a object}

\item{nomvarqual}{name of var}
}
\value{

}
\description{
TValueTable.FactorDataset
}
