% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumDataset.R
\name{TValueTable.NumDataset}
\alias{TValueTable.NumDataset}
\title{TValueTable.NumDataset}
\usage{
TValueTable.NumDataset(object)
}
\arguments{
\item{object}{a object NumData}
}
\value{

}
\description{
TValueTable.NumDataset
}
