% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etude_quali.R
\name{TValueTable.UnivFactor}
\alias{TValueTable.UnivFactor}
\title{TValueTable}
\usage{
TValueTable.UnivFactor(object)
}
\arguments{
\item{object}{a UnivFactor object}
}
\value{
a matrice
}
\description{
TValueTable
}
\examples{

}
