% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etude_quali.R
\name{UnivConst}
\alias{UnivConst}
\title{UnivConst}
\usage{
UnivConst(data, varqual1, vargroupe)
}
\arguments{
\item{data}{a data frame}

\item{varqual1}{a name variable}

\item{vargroupe}{a name variable}
}
\value{
a matrice
}
\description{
UnivConst
}
