% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{Vcramer.Data}
\alias{Vcramer.Data}
\title{Vcramer.Data}
\usage{
Vcramer.Data(object, var = FALSE)
}
\arguments{
\item{object}{a dataset object}

\item{var}{If you want the cramer value of an unique category variable}
}
\value{

}
\description{
Vcramer.Data
}
