% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor.R
\name{Vcramer.MultiFactor}
\alias{Vcramer.MultiFactor}
\title{Vcramer.UnivFactor}
\usage{
Vcramer.MultiFactor(object)
}
\arguments{
\item{object}{an UnivFator object}
}
\value{
a matrice
}
\description{
Vcramer.UnivFactor
}
