% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{vcramer}
\alias{vcramer}
\title{Cramer Values}
\usage{
vcramer(object, var = FALSE)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active category variable}
}
\value{
If you choose a specific variable, it's return a short sentence with the cramer value.
If you don't choose a specif variable, it's return a matrix with all the category variables of your dataset in lines with the cramer values in the first column.
}
\description{
Calul the cramer values between two category variables
}
