% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etude_quali.R
\name{Vcramer.UnivFactor}
\alias{Vcramer.UnivFactor}
\title{Vcramer.UnivFactor}
\usage{
Vcramer.UnivFactor(object)
}
\arguments{
\item{object}{an UnivFactor object}
}
\value{

}
\description{
Vcramer.UnivFactor
}
