% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{Visualisation.UnivFactor}
\alias{Visualisation.UnivFactor}
\title{Visualisation.UnivFactor}
\usage{
Visualisation.UnivFactor(object)
}
\arguments{
\item{object}{a UnivFactor object}
}
\value{
some plot
}
\description{
Visualisation.UnivFactor
}
