% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{VisualisationAC.Data}
\alias{VisualisationAC.Data}
\title{VisualisationAC.Data}
\usage{
VisualisationAC.Data(object, nomvarqual)
}
\arguments{
\item{object}{a object}

\item{nomvarqual}{a string}
}
\value{

}
\description{
VisualisationAC.Data
}
