% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor.R
\name{VisualisationAC.FactorDataset}
\alias{VisualisationAC.FactorDataset}
\title{VisualisationAC.FactorDataset}
\usage{
VisualisationAC.FactorDataset(object, nomvarqual)
}
\arguments{
\item{object}{a object}

\item{nomvarqual}{a string}
}
\value{

}
\description{
VisualisationAC.FactorDataset
}
