% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{VisualisationACM.Data}
\alias{VisualisationACM.Data}
\title{VisualisationACM.Data}
\usage{
VisualisationACM.Data(object)
}
\arguments{
\item{object}{a object}
}
\value{
a graph
}
\description{
VisualisationACM.Data
}
