% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor.R
\name{VisualisationACM.FactorDataset}
\alias{VisualisationACM.FactorDataset}
\title{VisualisationACM.FactorDataset}
\usage{
VisualisationACM.FactorDataset(object)
}
\arguments{
\item{object}{a object}
}
\value{
a graph
}
\description{
VisualisationACM.FactorDataset
}
