% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etude_quali.R
\name{VisualisationACM.UnivFactor}
\alias{VisualisationACM.UnivFactor}
\title{Visualisation.UnivFactor}
\usage{
VisualisationACM.UnivFactor(object)
}
\arguments{
\item{object}{a UnivFactor object}
}
\value{
some plot
}
\description{
Visualisation.UnivFactor
}
