% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor.R
\name{VisualisationAFC.MultiFactor}
\alias{VisualisationAFC.MultiFactor}
\title{VisualisationAFC.MultiFactor}
\usage{
VisualisationAFC.MultiFactor(object)
}
\arguments{
\item{object}{a object}
}
\value{
a graph
}
\description{
VisualisationAFC.MultiFactor
}
