% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{vizAFC}
\alias{vizAFC}
\title{AFC visualisation}
\usage{
vizAFC(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active variable}
}
\value{
A stack barplot with cluster in abscissa and modalities in stack.
A CA factor map which compare the cluster class and the modalities of the chosen variable.
}
\description{
Two graphs to illustrate the cluster with a specific category chosen variable.
}
