% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructeur.R
\name{Dataset}
\alias{Dataset}
\title{Constructor}
\usage{
Dataset(data, vargroupe)
}
\arguments{
\item{data}{a dataset that contains all the active variables}

\item{vargroupe}{a vector corresponding to the dataset clustering results}
}
\value{
an object returning a list of active data (categorical and numerical data), number of samples, etc.
}
\description{
Constructor
}
