% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructeur.R
\name{Dataset}
\alias{Dataset}
\title{Constructeur}
\usage{
Dataset(data, vargroupe)
}
\arguments{
\item{data}{a dataset (with all your active variables)}

\item{vargroupe}{the vector of your cluster}
}
\value{
an object which return a list of active data, category and numeric data, number of samples, etc.
}
\description{
Constructeur
}
