% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructeur.R
\name{Dataset}
\alias{Dataset}
\title{Constructor}
\usage{
Dataset(data, pred_clusters, true_clusters = FALSE)
}
\arguments{
\item{data}{a dataset that contains all the active variables}

\item{pred_clusters}{a vector of dataset predicted clusters}

\item{true_clusters}{a vector of dataset actual clusters}
}
\value{
an object returning a list of active data (categorical and numerical data), number of samples, etc.
}
\description{
Constructor
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
}
