% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{davies_bouldin}
\alias{davies_bouldin}
\title{davies_bouldin}
\usage{
davies_bouldin(data, clusters)
}
\arguments{
\item{data}{an object of class ccdata}

\item{clusters}{a vector corresponding to the dataset clustering results}
}
\value{

}
\description{
davies_bouldin
}
