% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{dunn_index}
\alias{dunn_index}
\title{dunn_index}
\usage{
dunn_index(data, clusters)
}
\arguments{
\item{data}{an object of class ccdata}

\item{clusters}{a vector corresponding to the dataset clustering results}
}
\value{

}
\description{
dunn_index
}
