% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFactor.R
\name{FactoDataset}
\alias{FactoDataset}
\title{MultiFactorConst}
\usage{
FactoDataset(data, vargroupe)
}
\arguments{
\item{data}{a dataset}

\item{vargroupe}{a string}
}
\value{
a object
}
\description{
MultiFactorConst
}
