% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{get_FAMD}
\alias{get_FAMD}
\title{Factorial Analysis of Mixed Data}
\usage{
get_FAMD(object)
}
\arguments{
\item{object}{An object of class ccdata}
}
\value{
three graphs to illustrate the cluster with all the variables of your dataset.
A. Plot with all the individuals colored by their cluster group and illustrated by categorical and numerical variables
B. Plot with all the modes of the categorical variables placed by importance of contribution with the first two axes.
C. Correlation circle with all the numerical variables
}
\description{
Factorial Analysis of Mixed Data
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
get_FAMD(obj)
}
