% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{get_MCA}
\alias{get_MCA}
\title{MCA}
\usage{
get_MCA(object, index_names)
}
\arguments{
\item{object}{An object of class ccdata}

\item{index_names}{Optional vector of the index to use

Show two graphs to illustrate the cluster with all the category variables of your dataset.}
}
\value{
A first plot with all the modalities of the category variables placed by importance of contribution with the two first axes.
A biplot whho illustrate the cluster group with all the modalities of category variables.
}
\description{
MCA
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
get_MCA(obj)
}
