% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComponentAnalysis.R
\name{get_PCA}
\alias{get_PCA}
\title{PCA visualisation}
\usage{
get_PCA(object)
}
\arguments{
\item{object}{an object of class ccdata}
}
\value{
Displays three graphs to allow a synthetic view of the data
A. Represents the correlation chart between all the variables
B. split into two graphs. On the left: the correlation circle and on the right: the biplot of the individuals and variables grouped by their clusters
}
\description{
Applying a principal component analysis (PCA) to the feature variables
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
get_PCA(obj)
}
