% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{phivalue}
\alias{phivalue}
\title{Value Phi}
\usage{
phivalue(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active category variable}
}
\value{
A cross table between the cluster variable and the chosen variable with phi values for all cluster*modality pairs.
}
\description{
Calcul the Phi value for all cluster*modalities of one chosen variable.
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
phivalue(obj, BankCustomer$profession)
}
