% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{PhiValueTable.Data}
\alias{PhiValueTable.Data}
\title{PhiValueTable.Data}
\usage{
PhiValueTable.Data(object, nomvarqual)
}
\arguments{
\item{object}{a object}

\item{nomvarqual}{a string}
}
\value{

}
\description{
PhiValueTable.Data
}
