% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotcorr}
\alias{plotcorr}
\title{Plot of correlations (numerical variables)}
\usage{
plotcorr(object)
}
\arguments{
\item{object}{An object of class ccdata}
}
\value{
A bar plot of the correlations between numerical variables and the cluster vector.
}
\description{
Plot of correlations (numerical variables)
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
plotcorr(obj)
}
