% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotphi}
\alias{plotphi}
\title{Title}
\usage{
plotphi(object, var)
}
\arguments{
\item{object}{an object of class ccdata}

\item{var}{a data vector of an active variable}
}
\value{

}
\description{
Title
}
