% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotphi}
\alias{plotphi}
\title{Plot of Phi values (categorical variables)}
\usage{
plotphi(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active variable}
}
\value{
A plot for all the phi values between cluster group and modalities of the chosen variables.
}
\description{
Plot of Phi values (categorical variables)
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
plotphi(obj, BankCustomer$profession)
}
