% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotsizeeff}
\alias{plotsizeeff}
\title{Plot of size effect (numerical variables)}
\usage{
plotsizeeff(object)
}
\arguments{
\item{object}{An object of class ccdata}
}
\value{
A plot of the effect size for all numerical variables by cluster.
}
\description{
Plot of size effect (numerical variables)
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
plotsizeeff(obj)
}
