% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plottvalue}
\alias{plottvalue}
\title{plottvalue}
\usage{
plottvalue(object, var = NULL)
}
\arguments{
\item{object}{an object of class ccdata}

\item{var}{a data vector of an active variable}
}
\value{

}
\description{
plottvalue
}
