% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotVCramer}
\alias{plotVCramer}
\title{plotVCramer}
\usage{
plotVCramer(object, limit = 10)
}
\arguments{
\item{object}{an object of class ccdata}

\item{limit}{number of categorical variables to display by descending value (default=10)}
}
\value{

}
\description{
plotVCramer
}
