% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eval.R
\name{plotVCramer}
\alias{plotVCramer}
\title{Plot of Cramer's V (categorical variables)}
\usage{
plotVCramer(object, limit = 10)
}
\arguments{
\item{object}{An object of class ccdata}

\item{limit}{Number of categorical variables to display by descending value (default=10)}
}
\value{
A bar plot with the Cramer values between the categorical variables and the cluster vector.
}
\description{
Plot of Cramer's V (categorical variables)
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
plotVCramer(obj)
}
