% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{statistical_test}
\alias{statistical_test}
\title{statistical_test}
\usage{
statistical_test(object, varexp)
}
\arguments{
\item{object}{an object of class ccdata}

\item{varexp}{a string}
}
\value{

}
\description{
statistical_test
}
