% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{statistical_test}
\alias{statistical_test}
\title{Statistical Test}
\usage{
statistical_test(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{Vector of a variable present in the dataset}
}
\value{
A sentence indicating whether the two variables are significantly different
}
\description{
Statistical Test
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
statistical_test(obj,BankCustomer$profession) #the cluster have significantly the same job?
statistical_test(obj,BankCustomer$revenu) #The cluster have significantly the same salary?
}
