% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{TValueTable.Data}
\alias{TValueTable.Data}
\title{TValueTable.Data}
\usage{
TValueTable.Data(object, varqual)
}
\arguments{
\item{object}{a data object}

\item{varqual}{vector of the cluster}
}
\value{

}
\description{
TValueTable.Data
}
