% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{validation}
\alias{validation}
\title{Evaluates the performance of the classifier by comparing predicted vs true clusters (true clusters required as input)}
\usage{
validation(object, true_clusters = object$true_clusters)
}
\arguments{
\item{object}{An object of class ccdata}

\item{true_clusters}{Vector of the true clusters (true clusters with ccdata class object by default)}
}
\value{
Confusion matrix, error rate, recall and precision. A matrix of recall and precision for all the cluster groups.
}
\description{
Evaluates the performance of the classifier by comparing predicted vs true clusters (true clusters required as input)
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
validation(obj, BankCustomer$Cluster)
}
