% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDataset.R
\name{Vcramer.Data}
\alias{Vcramer.Data}
\title{Vcramer.Data}
\usage{
Vcramer.Data(object, var = FALSE)
}
\arguments{
\item{object}{a object}

\item{var}{if you just want the Cramer for one varible}
}
\value{

}
\description{
Vcramer.Data
}
