% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateFactor.R
\name{vcramer}
\alias{vcramer}
\title{Cramer Values}
\usage{
vcramer(object, var = FALSE)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active categorical variable}
}
\value{
table of Cramer'V values for all categorical variables or only selected one
}
\description{
Calulates the cramer's V values between two categorical variables
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
vcramer(obj) #for all the cramer value by variables
vcramer(obj, BankCustomer$profession) #For one categorical variable
}
