% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateFactor.R
\name{vizAFC}
\alias{vizAFC}
\title{Correspondence Analysis visualisation
(two graphs to illustrate the cluster with a specific categorical selected variable)}
\usage{
vizAFC(object, var)
}
\arguments{
\item{object}{An object of class ccdata}

\item{var}{A data vector of an active variable}
}
\value{
A stack barplot with cluster in abscissa and modes in stack and a CA factor map which compare the cluster class and the modes of the selected variable.
}
\description{
Correspondence Analysis visualisation
(two graphs to illustrate the cluster with a specific categorical selected variable)
}
\examples{
data(BankCustomer)
obj <- Dataset(BankCustomer, BankCustomer$Cluster)
vizAFC(obj, BankCustomer$profession)
}
